requirements.txt
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-html>=3.2.0
requests>=2.31.0
pydantic>=2.0.0
python-dotenv>=1.0.0
colorama>=0.4.6
tabulate>=0.9.0 